Court-Data Fetcher & Mini-Dashboard/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
PyPDF2==3.0.1
lxml==4.9.3
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
urllib3==2.0.7
soupsieve==2.5
trio==0.23.1
trio-websocket==0.11.1
outcome==1.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
attrs==23.1.0
h11==0.14.0
wsproto==1.2.0
exceptiongroup==1.1.3
packaging==23.2
platformdirs==3.11.0
tqdm==4.66.1